read correct bytes to obtain ipv4 client address
